requirements.txt
line-bot-sdk
flask
gunicorn
pymongo
beautifulsoup4
requests